[api/ui] Added member archiving (closes #31)

+++ app/Providers/AppServiceProvider.php
@@ -34,16 +34,23 @@ class AppServiceProvider extends ServiceProvider
                 Route::get($name . "/reload", [$controller, "reload"])->middleware('permission:' . $name . '.reload');
             }
 
-            if (!in_array('final', $options)) {
+            if (in_array('archive', $options)) {
+                Route::get($name . "/archive", [$controller, "archive"])->middleware('permission:' . $name . '.archive');
+            }
+
+            if (!in_array('final', $options) && !in_array('readonly', $options)) {
                 Route::delete($name . "/{id}", [$controller, "destroy"])->middleware('permission:' . $name . '.delete');
                 Route::delete($name, [$controller, "destroyMany"])->middleware('permission:' . $name . '.delete');
                 Route::put($name . "/{id}", [$controller, "update"])->middleware('permission:' . $name . '.update');
                 Route::put($name, [$controller, "updateMany"])->middleware('permission:' . $name . '.update');
             }
 
+            if (!in_array('readonly', $options)) {
+                Route::post($name, [$controller, "store"])->middleware('permission:' . $name . '.create');
+            }
+
             Route::get($name, [$controller, "index"])->middleware('permission:' . $name . '.show');
             Route::get($name . "/{id}", [$controller, "show"])->middleware('permission:' . $name . '.show');
-            Route::post($name, [$controller, "store"])->middleware('permission:' . $name . '.create');
         });
     }
 }
